requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
h5py==3.11.0
importlib-resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pycryptodome==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.13.0
six==1.16.0
zipp==3.18.1
